# docs/swagger.yaml
openapi: 3.0.1
info:
  title: RadElement
  version: v1
paths:
...
  "/radelement/api/v1/sets":
    get:
      tags:
      - ElementSet
      summary: Fetches all the sets.
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
    post:
      tags:
      - ElementSet
      summary: Creates a set.
      requestBody:
        description: The content.
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateUpdateSet"
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
  "/radelement/api/v1/sets/{setId}":
    get:
      tags:
      - ElementSet
      summary: Fetches the set by set identifier.
      parameters:
      - name: setId
        in: path
        description: The set identifier.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
    put:
      tags:
      - ElementSet
      summary: Updates a set by set identifier.
      parameters:
      - name: setId
        in: path
        description: The set identifier.
        required: true
        schema:
          type: string
      requestBody:
        description: The content.
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateUpdateSet"
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
    delete:
      tags:
      - ElementSet
      summary: Removes a set by set identifier.
      parameters:
      - name: setId
        in: path
        description: The set identifier.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
  "/radelement/api/v1/sets/search":
    get:
      tags:
      - ElementSet
      summary: Searches the set with provided keyword.
      parameters:
      - name: searchKeyword
        in: query
        description: The search keyword.
        schema:
          type: string
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
...
  "/radelement/api/v1/organizations":
    get:
      tags:
      - Organization
      summary: Gets the organizations.
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
    post:
      tags:
      - Organization
      summary: Creates the organization.
      requestBody:
        description: The organization.
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateUpdateOrganization"
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
  "/radelement/api/v1/organizations/{organizationId}":
    get:
      tags:
      - Organization
      summary: Gets the organization.
      parameters:
      - name: organizationId
        in: path
        description: The organization identifier.
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
    put:
      tags:
      - Organization
      summary: Updates the organization.
      parameters:
      - name: organizationId
        in: path
        description: The organization identifier.
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        description: The organization.
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateUpdateOrganization"
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
    delete:
      tags:
      - Organization
      summary: Deletes the organization.
      parameters:
      - name: organizationId
        in: path
        description: The organization identifier.
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
  "/radelement/api/v1/organizations/search":
    get:
      tags:
      - Organization
      summary: Searches the organizations.
      parameters:
      - name: searchKeyword
        in: query
        description: The search keyword.
        schema:
          type: string
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - Bearer:
        - UserIdExists
...
components:
  schemas:
...
    ModalityType:
      enum:
      - CR
      - CT
      - DX
      - IVUS
      - MG
      - MR
      - NM
      - PT
      - RF
      - RG
      - US
      - XA
      type: string
    BiologicalSex:
      enum:
      - M
      - F
      type: string
...
    PersonRole:
      enum:
      - Author
      - Editor
      - Translator
      - Reviewer
      - Contributor
      type: string
    PersonDetails:
      type: object
      properties:
        personId:
          type: integer
          format: int32
        roles:
          type: array
          items:
            "$ref": "#/components/schemas/PersonRole"
          nullable: true
      additionalProperties: false
    OrganizationRole:
      enum:
      - Author
      - Sponsor
      - Translator
      - Reviewer
      - Contributor
      type: string
    OrganizationDetails:
      type: object
      properties:
        organizationId:
          type: integer
          format: int32
        roles:
          type: array
          items:
            "$ref": "#/components/schemas/OrganizationRole"
          nullable: true
      additionalProperties: false
    SpecialtyValue:
      type: object
      properties:
        value:
          maxLength: 2
          type: string
          nullable: true
      additionalProperties: false
...
    CreateUpdateSet:
      required:
      - contactName
      - description
      - name
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        contactName:
          type: string
        parentId:
          type: integer
          format: int32
          nullable: true
        modality:
          type: array
          items:
            "$ref": "#/components/schemas/ModalityType"
          nullable: true
        biologicalSex:
          type: array
          items:
            "$ref": "#/components/schemas/BiologicalSex"
          nullable: true
        ageUpperBound:
          type: number
          format: float
          nullable: true
        ageLowerBound:
          type: number
          format: float
          nullable: true
        version:
          maxLength: 8
          type: string
          nullable: true
        versionDate:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        persons:
          type: array
          items:
            "$ref": "#/components/schemas/PersonDetails"
          nullable: true
        organizations:
          type: array
          items:
            "$ref": "#/components/schemas/OrganizationDetails"
          nullable: true
        indexCodeReferences:
          type: array
          items:
            type: integer
            format: int32
          nullable: true
        referencesRef:
          type: array
          items:
            type: integer
            format: int32
          nullable: true
        specialties:
          type: array
          items:
            "$ref": "#/components/schemas/SpecialtyValue"
          nullable: true
      additionalProperties: false
...
    CreateUpdateOrganization:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 255
          type: string
        abbreviation:
          maxLength: 255
          type: string
          nullable: true
        url:
          maxLength: 255
          type: string
          nullable: true
        comment:
          maxLength: 255
          type: string
          nullable: true
        twitterHandle:
          maxLength: 255
          type: string
          nullable: true
        email:
          maxLength: 255
          type: string
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: 'JWT Authorization header. Example: {token}'
      scheme: bearer
      bearerFormat: JWT